refactor(build): Scope component selection to in-tree builds
