+++ .github/workflows/check-generate-otel-collector-distro.yml
@@ -37,7 +37,7 @@ jobs:
         run: |
           make generate-component-selectors generate-otel-collector-distro
 
-      - name: Test lightweight builders and component selectors
+      - name: Test lightweight builds and component selectors
         run: make test-lightweight-components
 
       - name: Check for generated file changes
