--- collector/testdata/external/minimal-builder-config.yaml
@@ -1,32 +1,0 @@
-dist:
-  module: example.com/minimal-alloy-collector
-  name: minimal-alloy-collector
-  description: Minimal external Alloy-aware OCB fixture.
-  version: v0.0.0
-  output_path: ./build/test-alloy-builder-external
-  build_tags: fixture_user_tag
-  cgo_enabled: true
-
-extensions:
-  - gomod: github.com/grafana/alloy v0.0.0
-    import: github.com/grafana/alloy/extension/alloyengine
-    name: alloyengine
-
-providers:
-  - gomod: go.opentelemetry.io/collector/confmap/provider/fileprovider v1.64.0
-  - gomod: go.opentelemetry.io/collector/confmap/provider/yamlprovider v1.64.0
-
-alloy_components:
-  - loki.write
-  - remote.http
-
-replaces:
-  # The fixture deliberately builds against this checkout rather than relying
-  # on an Alloy release that already contains cmd/alloy-builder.
-  - github.com/grafana/alloy => ../..
-  - github.com/grafana/alloy/syntax => ../../syntax
-  # The root module uses a zero placeholder and resolves it through this fork.
-  # OCB-generated modules must carry the replacement explicitly.
-  - go.opentelemetry.io/ebpf-profiler => github.com/grafana/opentelemetry-ebpf-profiler v0.0.202602-0.20260724100839-b73e704d1c59
-  # Alloy uses encoder hooks provided by this maintained YAML fork.
-  - gopkg.in/yaml.v2 => github.com/rfratto/go-yaml v0.0.0-20211119180816-77389c3526dc
