+++ internal/cmd/alloy-build-plan/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "os"
+
+func main() {
+	os.Exit(runBuildPlan(os.Args[1:], os.Stdin, os.Stdout, os.Stderr, os.Environ(), execDelegate{}))
+}
